backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
celery==5.5.3
redis==6.2.0
python-telegram-bot==22.3
openpyxl==3.1.5
Pillow==11.3.0
django-extensions==4.1
